+++ crates/evpn-linux/src/linux/mod.rs
@@ -709,6 +709,8 @@ impl Dataplane for LinuxDataplane {
             | DataplaneOp::RemoveManagedBridge { .. }
             | DataplaneOp::CreateManagedVxlan { .. }
             | DataplaneOp::RemoveManagedVxlan { .. }
+            | DataplaneOp::CreateManagedSvdVxlan { .. }
+            | DataplaneOp::RemoveManagedSvdVxlan { .. }
             | DataplaneOp::CreateManagedVrf { .. }
             | DataplaneOp::RemoveManagedVrf { .. }
             | DataplaneOp::CreateManagedL3Vxlan { .. }